document: uses extensions
* Uses record extensions to dump additional information such as
  concatenate text version.
