+++ rero_ils/modules/acquisition/acq_order_lines/dumpers.py
@@ -26,7 +26,7 @@ from rero_ils.modules.acquisition.acq_order_lines.models import \
     AcqOrderLineNoteType
 from rero_ils.modules.commons.identifiers import IdentifierType
 from rero_ils.modules.documents.dumpers import DocumentAcquisitionDumper
-from rero_ils.modules.documents.utils import title_format_text_head
+from rero_ils.modules.documents.extensions import TitleExtension
 
 
 class AcqOrderLineESDumper(InvenioRecordsDumper):
@@ -69,7 +69,7 @@ class AcqOrderLineESDumper(InvenioRecordsDumper):
 
         data['document'] = {
             'pid': document.pid,
-            'title': title_format_text_head(document.get('title', [])),
+            'title': TitleExtension.format_text(document.get('title', [])),
             'identifiers': identifiers
         }
         data['document'] = {k: v for k, v in data['document'].items() if v}
